ship builtins and helpers as separate modules
Prepending only the necessary ones to the output was a neat idea
but didn't work in practice due to function parameters aliasing.

To make it work as is the sources would either have to be parsed and
any aliasing identifiers mangled or you could just not prepend them
to the output.