rework `std.Build` usage/integration
This commit effectively rewrites `build.zig`, `build/TranslateC.zig`
(now `build/Translator.zig`), and the example build scripts.

Instead of being its own step, `Translator` is a helper type which
simply wraps a `Build.Step.Run` and an associated `*Build.Module`. This
is not only simpler, but also necessary, because custom steps are going
to be removed from `std.Build` at some point. Many of the same APIs as
before are available, but also several more. For instance, you can call
`Translator.linkLibrary` to both link a library (as a `Step.Compile`) to
the generated `*Build.Module`, *and* expose its installed headers to
`translate-c`. In general, any method of `*Build.Module` which can add
an include path or similar is replicated on `Translator`.

The `build.zig` has been improved and simplified. The old script was
quite overcomplicated, and also relied on some hacks and even
technically on some Illegal Behavior. Much of this complexity arose from
the fact that the script was trying to make `zig build test` run the
translate-c tests in *all* optimization modes. This is unnecessary
effort; it's much simpler to just run `zig build` a few separate times
from the CI workflow.

The `test-examples` step was previously disabled due to a Zig compiler
bug with the recursive imports (`build.zig` depends on
`examples/x/build.zig` which depends on `build.zig` again). While the
exact current output is a compiler bug, it's actually not possible for
recursive package dependencies to work; you would end up infinitely
recursing through `fn build` calls. So, the `test-examples` step is
removed. It is replaced with a new `examples/build.zig`, which just
aggregates all of the examples so you can run `zig build all` in that
directory to test them all. The CI workflow now runs this command too.

Lastly, the examples are refactored to use the new APIs, be a bit
clearer, and to properly test their output (`import_header` by actually
running the test it builds, `compile_c` by checking the captured stdout
of both executables after running them).

Oh, and I added a new example, `use_static_lib`, demonstrating the
ability of the new API to completely wrap a C library (by translating
its headers, but also including its source in the generated module) in a
concise way.