examples: handle Windows CRT stdout line endings
On Windows, the CRT (libc) will by default have stdio streams open in
"text" mode, so will convert "\n" into "\r\n" in output. We could
change the file mode in the examples behind a `#ifdef _WIN32`, but it's
a bit easier to just handle the difference in the build script checks.