ci: test against `minimum_zig_version`
mlugg/setup-zig v2 will do this by default if the version isn't
specified in the manifest, and it seems like a better idea here. We're
keeping `minimum_zig_version` up-to-date anyway; this change decreases
maintenance burder if and when we pin to full Zig releases, and it means
that if there is a breaking change upstream, our CI doesn't break.