build: introduce `-Dtest-cross-targets`
This new `zig build` option causes the cases to be translated not only
for the `-Dtarget` (typically `native`) target, but also (unless they
have an explicit target override) a list of 5 cross targets:

* `x86_64-linux`
* `x86_64-windows`
* `aarch64-linux`
* `aarch64-macos`
* `arm-linux`

The CI workflow is updated to test these cross targets (in Debug mode).

Almost every existing case was able to pass on these cross targets
without any changes. Aro was missing support for varargs on ARM (see
Vexu/arocc#869) which caused any test importing `stdio.h` to fail on
`arm-linux`, and one of the run cases needed an `#ifdef` added to avoid
assuming `__int128`, but that's all!